+++ src/pages/CompaniesPage/ExperiencePage/ExperiencePage.tsx
@@ -1,4 +1,4 @@
-import ExperienceCard from '@/components/common/ExperienceCard';
+import ExperienceCard from '@/components/common/ExperienceCard/ExperienceCard';
 import { Button } from '@/components/ui/button';
 import {
   DropdownMenu,
